ISSUE #4224 - new group works
