+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsForm/ticketGroups/groupsAccordion/groupsAccordion.styles.ts
@@ -19,6 +19,7 @@ import styled, { css } from 'styled-components';
 import { Accordion as AccordionBase } from '@controls/accordion/accordion.component';
 import { CentredContainer } from '@controls/centredContainer';
 import CheckBoxBase from '@mui/material/Checkbox';
+import { Button } from '@controls/button';
 import { CollectionAccordion } from '../groups/groups.styles';
 import { Container as GroupItemContainer } from '../groups/groupItem/groupItem.styles';
 
@@ -93,4 +94,12 @@ export const NumberContainer = styled(CentredContainer)`
 
 export const Checkbox = styled(CheckBoxBase)`
 	margin: 0 8px 0 auto;
+`;
+
+export const NewGroupButton = styled(Button).attrs({
+	variant: 'outlined',
+	color: 'secondary',
+})`
+	align-self: flex-end;
+	margin: 14px 0 3px;
 `;
